feat(db): add project_limits, an override the limits owner can write
team_limits derives a team's limits by joining tiers to addons. addons is
written only by billing and read by nothing in this repo except that
view, so the sandbox-creation read path depends on a table this side does
not own -- and there is nowhere for an owner to push an effective limit
to, which is what blocks upsertProjectLimits on the management API.

project_limits is that destination. Every view column becomes
COALESCE(override, existing expression), so an empty table changes
nothing. Nothing writes it yet, and landing it alone is the point: it
edits the object gating every sandbox creation while nothing depends on
the edit.

All nine columns are NOT NULL, so a team is overridden completely or not
at all; there is no half-overridden row to reason about. They are bigint
to match what the view already yields, and CHECK >= 0 rather than tiers'
> 0 -- this is a push target, and rejecting a value the caller considers
valid turns a product decision into a retry loop that never drains.
Negatives are always a bug and stay rejected.

The generated TeamLimit widens from int32 to int64 on six columns. That
is a fix, not a consequence: the columns are bigint, and sqlc could not
previously infer the type of the view's arithmetic so it defaulted to
int32. It resolves through COALESCE now. The one caller feeding an int32
response schema narrows explicitly at that boundary, where the contract
demands it, instead of relying on codegen to do it invisibly.

Tests prove the three properties the migration claims. Inertness is
asserted against the arithmetic the view used to perform rather than
fixed numbers, so it stays honest if tier defaults move. Both it and the
override test are mutation-verified: dropping an addon term from one
column, and reversing a COALESCE, each fail the intended test.

+++ packages/db/pkg/auth/queries/team_creation_guard.sql.go
@@ -56,12 +56,12 @@ type GetTeamsWithUsersTeamsWithTierForUpdateRow struct {
 	IsDefault                bool
 	ID_2                     uuid.UUID
 	MaxLengthHours           int64
-	ConcurrentSandboxes      int32
-	ConcurrentTemplateBuilds int32
-	MaxVcpu                  int32
-	MaxRamMb                 int32
-	DiskMb                   int32
-	EventsTtlDays            int32
+	ConcurrentSandboxes      int64
+	ConcurrentTemplateBuilds int64
+	MaxVcpu                  int64
+	MaxRamMb                 int64
+	DiskMb                   int64
+	EventsTtlDays            int64
 }
 
 func (q *Queries) GetTeamsWithUsersTeamsWithTierForUpdate(ctx context.Context, userID uuid.UUID) ([]GetTeamsWithUsersTeamsWithTierForUpdateRow, error) {
